Moves parsing utilities one level up to parsing. Gets rid of parsing/sdf
