+++ drake/examples/double_pendulum/BUILD.bazel
@@ -20,7 +20,7 @@ drake_cc_library(
     deps = [
         "//drake/common:essential",
         "//drake/multibody:rigid_body_tree",
-        "//drake/multibody/parsing/sdf:frame_cache",
+        "//drake/multibody/parsing:frame_cache",
         "@eigen",
         "@sdformat",
     ],
