+++ drake/multibody/parsing/frame_cache.h
